g_red_brick: Properly handle Microsoft Extended Compat ID and Properties Descriptors
Expose MSFT100* as a normal string descriptor on index 0xEE.

composite.c needs a hack to handle Windows asking for string descriptor
with language set to 0x0000. For the language to 0x0409 for the 0xEE
descriptor.

Make all the related structs packed, othewise the padding could destroy
the data layout.