g_red_brick: Fix USB request handling
Don't queue a request while the context spinlock is held, because
usb_eq_queue() might directly call the complete function which will
try to lock the context spinlock too.

This also means that the request has to be added to the active list
before it is queue, so that the complete function can move it back to
the idle list without creating list corruption.