red-brick: Inverted RTS assert and deassert logic.
This was needed because internally for the chip the RTS pin is active low. But
The chip on the RS485 extension has TXE pin as active high. With this modification
there is no need for an inverter to invert the active low signal to be used as
TXE pin input.